# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StruturaCheck - Structural Defect Detection System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: analyses
    description: Image analysis operations
  - name: stats
    description: Statistics and dashboard data
  - name: recommendations
    description: AI maintenance recommendations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Registration failed

  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List all analyses
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"

    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Create a new analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"

  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a single analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found

    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get dashboard statistics summary
      responses:
        "200":
          description: Statistics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/history:
    get:
      operationId: getStatsHistory
      tags: [stats]
      summary: Get defect detection history over time
      responses:
        "200":
          description: History data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryPoint"

  /stats/distribution:
    get:
      operationId: getDefectDistribution
      tags: [stats]
      summary: Get defect distribution by structure type
      responses:
        "200":
          description: Distribution data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DistributionItem"

  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List AI maintenance recommendations
      responses:
        "200":
          description: List of recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  /recommendations/{id}/work-order:
    post:
      operationId: generateWorkOrder
      tags: [recommendations]
      summary: Generate a work order for a recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work order generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"

  /recommendations/{id}/reminder:
    post:
      operationId: setInspectionReminder
      tags: [recommendations]
      summary: Set inspection reminder for a recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "200":
          description: Reminder set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"

  /reports/{analysisId}:
    get:
      operationId: downloadReport
      tags: [analyses]
      summary: Get report data for an analysis
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]

    Analysis:
      type: object
      required: [id, fileName, structureType, severity, status, createdAt]
      properties:
        id:
          type: integer
        fileName:
          type: string
        structureType:
          type: string
          description: "bridge, road, wall, building"
        severity:
          type: string
          description: "high, medium, low, none"
        status:
          type: string
          description: "pending, processing, completed, failed"
        defectCount:
          type: ["integer", "null"]
        confidenceScore:
          type: ["number", "null"]
        analysisSpeedMs:
          type: ["integer", "null"]
        originalImageUrl:
          type: ["string", "null"]
        maskedImageUrl:
          type: ["string", "null"]
        defectTypes:
          type: ["string", "null"]
          description: "JSON array of defect type strings"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    AnalysisInput:
      type: object
      required: [fileName, structureType]
      properties:
        fileName:
          type: string
        structureType:
          type: string
        notes:
          type: string

    StatsSummary:
      type: object
      required: [totalImagesChecked, totalDefectsFound, avgAnalysisSpeedMs, highSeverityCount, mediumSeverityCount, lowSeverityCount, trendPercent]
      properties:
        totalImagesChecked:
          type: integer
        totalDefectsFound:
          type: integer
        avgAnalysisSpeedMs:
          type: number
        highSeverityCount:
          type: integer
        mediumSeverityCount:
          type: integer
        lowSeverityCount:
          type: integer
        trendPercent:
          type: number

    HistoryPoint:
      type: object
      required: [date, defectsFound, imagesChecked]
      properties:
        date:
          type: string
        defectsFound:
          type: integer
        imagesChecked:
          type: integer

    DistributionItem:
      type: object
      required: [type, count, percentage]
      properties:
        type:
          type: string
        count:
          type: integer
        percentage:
          type: number

    Recommendation:
      type: object
      required: [id, analysisId, severity, title, description, reasoning, createdAt]
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        severity:
          type: string
          description: "critical, warning, safe"
        title:
          type: string
        description:
          type: string
        reasoning:
          type: string
          description: "Markdown content explaining the AI reasoning"
        workOrderGenerated:
          type: boolean
        reminderDate:
          type: ["string", "null"]
        createdAt:
          type: string

    WorkOrder:
      type: object
      required: [id, recommendationId, reference, createdAt]
      properties:
        id:
          type: integer
        recommendationId:
          type: integer
        reference:
          type: string
        createdAt:
          type: string

    ReminderInput:
      type: object
      required: [reminderDate]
      properties:
        reminderDate:
          type: string

    Report:
      type: object
      required: [analysisId, fileName, structureType, severity, createdAt, defectCount, recommendations]
      properties:
        analysisId:
          type: integer
        fileName:
          type: string
        structureType:
          type: string
        severity:
          type: string
        createdAt:
          type: string
        defectCount:
          type: integer
        confidenceScore:
          type: ["number", "null"]
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"