# lib/api-zod/src/generated/types/historyPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StruturaCheck - Structural Defect Detection System API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryPoint {
  date: string;
  defectsFound: number;
  imagesChecked: number;
}