# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StruturaCheck - Structural Defect Detection System API
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  id: number;
  analysisId: number;
  /** critical, warning, safe */
  severity: string;
  title: string;
  description: string;
  /** Markdown content explaining the AI reasoning */
  reasoning: string;
  workOrderGenerated?: boolean;
  /** @nullable */
  reminderDate?: string | null;
  createdAt: string;
}