# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StruturaCheck - Structural Defect Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { Recommendation } from './recommendation';

export interface Report {
  analysisId: number;
  fileName: string;
  structureType: string;
  severity: string;
  createdAt: string;
  defectCount: number;
  /** @nullable */
  confidenceScore?: number | null;
  recommendations: Recommendation[];
}