# lib/api-zod/src/generated/types/workOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StruturaCheck - Structural Defect Detection System API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkOrder {
  id: number;
  recommendationId: number;
  reference: string;
  createdAt: string;
}